move out previous Cutscene logic from WorldStateModel to CutsceneModel
