+++ common/src/main/java/com/wynntils/models/cutscene/CutsceneModel.java
@@ -4,10 +4,13 @@
  */
 package com.wynntils.models.cutscene;
 
+import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.components.Model;
 import com.wynntils.core.components.Models;
 import com.wynntils.mc.event.SetCameraEntityEvent;
 import com.wynntils.models.character.type.VehicleType;
+import com.wynntils.models.cutscene.type.CutsceneState;
+import com.wynntils.models.worlds.event.CutsceneStartedEvent;
 import java.util.List;
 import java.util.Optional;
 import net.minecraft.world.entity.Display;
@@ -16,6 +19,7 @@ import net.neoforged.bus.api.SubscribeEvent;
 
 public class CutsceneModel extends Model {
     private Optional<Entity> cameraEntity = Optional.empty();
+    private CutsceneState cutsceneState = CutsceneState.NOT_IN_CUTSCENE;
 
     public CutsceneModel() {
         super(List.of());
@@ -26,12 +30,30 @@ public class CutsceneModel extends Model {
         cameraEntity = Optional.ofNullable(event.getViewingEntity());
     }
 
+    public void cutsceneStarted(boolean groupCutscene) {
+        if (cutsceneState == CutsceneState.NOT_IN_CUTSCENE) {
+            cutsceneState = CutsceneState.IN_CUTSCENE;
+
+            CutsceneStartedEvent event = new CutsceneStartedEvent(groupCutscene);
+            WynntilsMod.postEvent(event);
+
+            if (event.isCanceled()) {
+                cutsceneState = CutsceneState.SKIPPED_CUTSCENE;
+            }
+        }
+    }
+
     public Optional<Entity> getCameraEntity() {
         return cameraEntity;
     }
 
     public boolean isCutsceneActive() {
         return Models.Character.getVehicle() == VehicleType.DISPLAY
-                || cameraEntity.filter(entity -> entity instanceof Display).isPresent();
+                || cameraEntity.filter(entity -> entity instanceof Display).isPresent()
+                || cutsceneState == CutsceneState.IN_CUTSCENE;
+    }
+
+    public void cutsceneEnded() {
+        cutsceneState = CutsceneState.NOT_IN_CUTSCENE;
     }
 }
