+++ common/src/main/java/com/wynntils/models/cutscene/type/CutsceneState.java
@@ -1,8 +1,8 @@
 /*
- * Copyright © Wynntils 2025.
+ * Copyright © Wynntils 2025-2026.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.worlds.type;
+package com.wynntils.models.cutscene.type;
 
 public enum CutsceneState {
     NOT_IN_CUTSCENE,
